Tasks are linked to Users, and CRUD of Tasks possible only by creator User of thoe tasks. Sorting, Pagination and Filtering of tasks also added.

+++ src/controllers/task.js
@@ -1,7 +1,11 @@
 const Task = require('../models/task');
 
 const createTask = async (req, res, next) => {
-    const task = new Task(req.body);
+    const task = new Task({
+        ...req.body, 
+        creator: req.user.id
+    });
+
     try {
         await task.save();
         return res.status(201).send({
@@ -16,10 +20,11 @@ const createTask = async (req, res, next) => {
     }
 };
 
+// Gets a task by the task ID:
 const getTask = async (req, res, next) => {
+    const taskId = req.params.id;
     try {
-        console.log(req.params.id);
-        const taskFound = await Task.findOne({ _id: req.params.id });
+        const taskFound = await Task.findOne({ _id: taskId });
         if (!taskFound)
             return res
                 .status(404)
@@ -32,6 +37,46 @@ const getTask = async (req, res, next) => {
             .send({ message: `Error occurred while fetching the task: ${e}` });
     }
 };
+
+// Gets a selection of task after filtering by query params, as follows:
+// FILTERING: GET /task?done=true (else done=false)
+// PAGINATION: GET /task?limit=10&page=1
+// SORTING: GET /task?sortBy=createdAt:desc
+const getSelectedTasks = async (req,res,next)=>{
+    
+    // Filtering todos which are 'done'.
+    const match ={};
+    if(req.query.done){
+        match.done = req.query.done === 'true';
+    }
+
+    // Sorting by createdAt: (or even other possible attributes).
+    const sort = {};
+    if (req.query.sortBy) {
+        const sortQuery = req.query.sortBy.split(':');
+        sort[sortQuery[0]] = sortQuery[1] === 'desc' ? -1 : 1;
+    }
+
+    // Pagination of TODDOs
+    const currentPage = Number(req.query.page) || 1;
+    const limitPerPage = Number(req.query.limit) || 4;
+    const skip = limitPerPage * (currentPage - 1);
+
+    try {
+        await req.user.populate({
+            path: 'tasks',
+            match,
+            options: {
+                limit: parseInt(limitPerPage),
+                skip: parseInt(skip),
+                sort,
+            },
+        });
+        res.send({tasks:req.user.tasks, message: 'Successfully fetched Tasks!'});
+    } catch (e) {
+        res.status(500).send({message:`Error occurred while fetching tasks. ${e}`});
+    }
+}
 
 const updateTask = async (req, res, next) => {
     // Check if all the updates which are asked to do, are even allowed:
@@ -48,7 +93,8 @@ const updateTask = async (req, res, next) => {
     }
 
     try {
-        const taskFound = await Task.findOne({ _id: req.params.id });
+        // Fetches a task only if the task id present in taskDB and also its creator is the user making request.
+        const taskFound = await Task.findOne({ _id: req.params.id, creator: req.user._id });
 
         if (!taskFound){
             return res
@@ -74,7 +120,7 @@ const updateTask = async (req, res, next) => {
 
 const deleteTask = async (req, res, next) => {
     try {
-        const taskFound = await Task.findOneAndDelete({ _id: req.params.id });
+        const taskFound = await Task.findOneAndDelete({ _id: req.params.id, creator: req.user._id });
 
         if (!taskFound) {
             return res.status(404).send({ message: 'Could not find the task' });
@@ -91,6 +137,7 @@ const deleteTask = async (req, res, next) => {
 module.exports = {
     createTask,
     getTask,
+    getSelectedTasks,
     updateTask,
     deleteTask,
 };
